Bug fixes, update the door code
* Properly do the timing code - I was running always once the time condition elapsed
+ Door code redone - can now show all doors currently open, using arrows
* Analysis setting doesn't create a pointer in mazda3BKLCDPrint, dereference the struct directly
* Logging will properly retain if it's been run once bfore
* Free the string created in stateToSerial
* Clarify if condition in binaryTreeFindR
* Use the same type in BinaryTreeNode and binaryTreeFindR

+++ canscan/naz-binaryTree.h
@@ -11,7 +11,7 @@ extern "C" {
 
 typedef struct BinaryTreeNode {
     struct BinaryTreeNode *parent;
-    uint16_t key;
+    int key;
     void *data;
     struct BinaryTreeNode *childLeft;
     struct BinaryTreeNode *childRight;
